refactor tests and benches
Separate tests and benchmarks, only enable feature(tests) for benchmarks.
